refactor(core/rust): use separate files for button and touch events
this makes it easier to feature-gate them
